+++ src/wai/annotations/isp/write_labels/specifier/__init__.py
@@ -1,0 +1,1 @@
+from ._WriteLabelsISPSpecifier import WriteLabelsISPSpecifier